Se agrega diagrama, y breve explicación de la resolución
