+++ oia/lanzamiento-de-aceitunas/Makefile
@@ -1,0 +1,2 @@
+uml:
+	java -jar ../../utils/plantuml.jar ./diagrama.plantuml
